+++ src/gort/recipes/calibration.py
@@ -267,7 +267,7 @@ class TwilightFlats(BaseRecipe):
                 # exposure time is now too long. Exit.
                 gort.log.debug("Exposure time is too long to take more extra flats.")
                 break
-            elif is_sunrise and all_done and exp_time <= self.MAX_EXP_TIME_EXTRA:
+            elif is_sunset and all_done and exp_time <= self.MAX_EXP_TIME_EXTRA:
                 # Continue taking exposures for extra flats.
                 pass
             else: