+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 🔧 Fixed
 
 * Prevent the Overwatcher observer from opening the dome while calibrations are ongoing.
+* Fixed a bug in the twilight flats recipe related to the extra exposures.
 
 
 ## 1.1.0 - November 7, 2024